mikado/requirements.txt
amqp==5.2.0
argcomplete==3.2.3
arrow==1.3.0
asgiref==3.8.1
astroid==3.1.0
asttokens==2.4.1
async-timeout==4.0.3
autopep8==2.1.0
Babel==2.14.0
billiard==4.2.0
binaryornot==0.4.4
black==24.3.0
celery==5.3.6
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
construct==2.5.3
cookiecutter==2.6.0
cron-descriptor==1.4.3
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.8
distro==1.9.0
Django==4.2.11
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-extra-views==0.14.0
django-haystack==3.2.1
django-localflavor==3.1
django-phonenumber-field==7.3.0
django-redis==5.3.0
django-tables2==2.7.0
django-timezone-field==6.1.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
flower==2.0.1
html5lib==1.1
humanize==4.9.0
idna==3.6
install==1.3.5
ipdb==0.13.13
ipython==8.18.1
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.3
kombu==5.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
netaddr==1.2.1
packaging==24.0
parso==0.8.3
pathspec==0.12.1
pefile==2023.2.7
phonenumbers==8.13.33
pillow==10.2.0
pipenv==2023.12.1
pipx==1.4.3
platformdirs==4.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pure-eval==0.2.2
purl==1.6
pycodestyle==2.11.1
Pygments==2.17.2
pylint==3.1.0
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyprof2calltree==1.4.5
pysolr==3.9.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-ptrace==0.9.9
python-slugify==8.0.4
python-stdnum==1.20
pytils==0.4.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
redis==5.0.3
regex==2024.4.28
requests==2.31.0
rich==13.7.1
setuptools==70.0.0
six==1.16.0
smsaero-api==2.2.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
stack-data==0.6.3
text-unidecode==1.3
textile==4.0.2
tomli==2.0.1
tomlkit==0.12.4
tornado==6.4.1
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
userpath==1.9.2
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.5.0
Whoosh==2.7.4
wrapt==1.16.0
yookassa==3.0.1
